% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{package}
\name{usemodels-package}
\alias{usemodels}
\alias{usemodels-package}
\title{usemodels: Boilerplate Code for 'Tidymodels' Analyses}
\description{
Code snippets to fit models using the tidymodels framework can be easily created for a given data set.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://usemodels.tidymodels.org/}
  \item \url{https://github.com/tidymodels/usemodels}
  \item Report bugs at \url{https://github.com/tidymodels/usemodels/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
